requirements.txt
Flask
Flask-SocketIO
PyMySQL
python-dotenv
pandas
joblib
numpy
gunicorn
scikit-learn
mysql-connector-python
matplotlib
seaborn 
